Docker image for Catch2
