+++ catch2/build_image.bat
@@ -1,0 +1,1 @@
+docker build --tag cw-catch2-cpp .
