+++ catch2/docker_image/.gitignore
@@ -1,0 +1,1 @@
+build/
