+++ catch2/docker_image/CMakeLists.txt
