+++ catch2/docker_image/configure.sh
@@ -1,2 +1,2 @@
-#!/bin/bash
+#!/bin/sh
 cmake -S . -B build -G Ninja
