+++ catch2/docker_image/run-kata.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+./configure.sh && ./run-tests.sh
