+++ catch2/docker_image/run-tests.sh
@@ -1,3 +1,3 @@
-#!/bin/bash
+#!/bin/sh
 
 cmake --build build && ./build/tests/tests --reporter=codewars
