+++ catch2/docker_image/tests/CMakeLists.txt
@@ -22,7 +22,7 @@ add_executable(tests codewars_reporter.cpp test_solution.cpp)
 target_compile_features(tests PRIVATE cxx_std_20)
 target_compile_options(tests PRIVATE -Wall -Wextra -O2)
 # target_link_libraries(tests PRIVATE challenge Catch2::Catch2 m crypto dl sqlite3 tbb fmt Threads::Threads)
-target_link_libraries(tests PRIVATE challenge Catch2::Catch2WithMain m crypto dl sqlite3 tbb fmt Threads::Threads)
+target_link_libraries(tests PRIVATE challenge Catch2::Catch2WithMain m dl fmt Threads::Threads)
 
 # Not really necessary for CR, but allows running tests through CMake
 add_test(NAME tests COMMAND tests)
