+++ catch2/kata_code/OriginalExample/tests/test_solution.cpp
@@ -1,6 +1,6 @@
 #include <vector>
 
-#include <challenge.h>
+#include <preloaded.h>
 #include <catch2/catch_test_macros.hpp>
