+++ catch2/run_image.bat
@@ -1,0 +1,8 @@
+set KATA=%1
+set IMAGE=cw-catch2-cpp
+set WORKDIR=/code/
+set CREATE_CMD=docker container create --rm -w %WORKDIR% %IMAGE% ./run-kata.sh
+
+FOR /F %%i IN ('%CREATE_CMD%') DO set CONTAINER=%%i
+docker cp ./kata_code/%KATA%/. %CONTAINER%:%WORKDIR%
+docker start --attach -i %CONTAINER%
